refactor:simplify the code structure
